refactor(mode): move AppMode enum into src/mode.rs
Repoint crate::app_mode::AppMode -> crate::mode::AppMode across the binary.
app_mode.rs now holds only the Default-mode UI lifecycle (moved next). No
behavior change.
