+++ src/mode/tmux/gate.rs
@@ -5,8 +5,8 @@
 //! `ozma_terminal`'s shared mouse systems yield to the tmux-specific gestures.
 
 use super::pane_hit::tmux_pane_at_phys;
-use crate::app_mode::AppMode;
 use crate::input::ime::ImeState;
+use crate::mode::AppMode;
 use crate::ui::copy_mode::CopyModeState;
 use crate::ui::copy_search::CopyPrompt;
 use crate::ui::rename_prompt::RenamePrompt;
