+++ src/window_title.rs
@@ -2,7 +2,7 @@
 //! into the primary window's title bar — `session:window — ozmux` in Tmux
 //! mode, the focused terminal's OSC title + ` — ozmux` in Default mode.
 
-use crate::app_mode::AppMode;
+use crate::mode::AppMode;
 use bevy::prelude::*;
 use bevy::window::{PrimaryWindow, Window};
 use ozma_terminal::{KeyboardFocused, OzmaTerminal};
